+++ src/GameJoltLibrary/GameJoltLibrary.cs
@@ -1,15 +1,8 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Linq;
-using System.Net.Http;
-using System.Threading;
 using System.Windows.Controls;
-using AngleSharp;
-using AngleSharp.Dom.Html;
-using GameJoltLibrary.Models;
-using Newtonsoft.Json;
 using Playnite.SDK;
-using Playnite.SDK.Data;
 using Playnite.SDK.Models;
 using Playnite.SDK.Plugins;
 
@@ -17,9 +10,9 @@ namespace GameJoltLibrary
 {
     public class GameJoltLibrary : LibraryPlugin
     {
-        private static readonly ILogger logger = LogManager.GetLogger();
+        private static readonly ILogger _logger = LogManager.GetLogger();
         public InstalledGamesProvider InstalledGamesProvider { get; }
-
+        public LibraryGamesProvider LibraryGamesProvider { get; }
         public GameJoltMetadataProvider MetadataProvider { get; }
 
         private GameJoltLibrarySettingsViewModel settingsViewModel { get; set; }
@@ -38,9 +31,10 @@ namespace GameJoltLibrary
             {
                 HasSettings = true
             };
-            logger.Info("GemeJolt library initialized.");
-            InstalledGamesProvider = new InstalledGamesProvider(logger);
-            MetadataProvider = new GameJoltMetadataProvider(this, logger);
+            _logger.Info("GemeJolt library initialized.");
+            InstalledGamesProvider = new InstalledGamesProvider(_logger);
+            LibraryGamesProvider = new LibraryGamesProvider(_logger);
+            MetadataProvider = new GameJoltMetadataProvider(this, _logger);
         }
 
         public override IEnumerable<GameMetadata> GetGames(LibraryGetGamesArgs args)
@@ -51,51 +45,12 @@ namespace GameJoltLibrary
             }
 
             var games = InstalledGamesProvider.GetInstalledGamesV2(args).Values.ToList();
-            var libraryGames = GetLibraryGames();
 
-            games.AddRange(libraryGames);
-            return games;
-        }
-
-        public IEnumerable<GameMetadata> GetLibraryGames()
-        {
-            var games = new List<GameMetadata>();
-
-            try
+            if (settingsViewModel.Settings.UserName is string userName)
             {
-                var settings = settingsViewModel.Settings;
-                string ownedGamesUrl = $"https://gamejolt.com/@{settings.UserName}/owned";
-
-                var http = new HttpClient();
-                http.BaseAddress = new Uri("https://gamejolt.com/site-api/");
-
-                var result = http.GetAsync("web/library/games/owned/@mrxx99").GetAwaiter().GetResult();
-
-                var stringContent = result.Content.ReadAsStringAsync().GetAwaiter().GetResult();
-
-                var ownedGames = Serialization.FromJsonStream<LibraryGamesResult>(result.Content.ReadAsStreamAsync().GetAwaiter().GetResult());
-
-                var config = Configuration.Default.WithDefaultLoader();
-                var context = BrowsingContext.New(config);
-
-                foreach (var ownedGame in ownedGames.Payload.Games)
-                {
-                    var game = new GameMetadata
-                    {
-                        GameId = ownedGame.Id.ToString(),
-                        Source = new MetadataNameProperty("Game Jolt"),
-                        Name = ownedGame.Title,
-                        IsInstalled = false,
-                        Links = new List<Link> { new Link("store page web", ownedGame.Link) },
-                        Developers = new HashSet<MetadataProperty> { new MetadataNameProperty(ownedGame.Developer.DisplayName) }
-                    };
-
-                    games.Add(game);
-                }
-            }
-            catch (Exception ex)
-            {
-                logger.Error(ex, "Error during loading library games for Game Jolt");
+                var libraryGames = LibraryGamesProvider.GetLibraryGames(userName);
+                var libraryGamesToAdd = libraryGames.Where(game => !games.Any(game => game.GameId == game.GameId));
+                games.AddRange(libraryGames);
             }
 
             return games;
